requirements.txt
python-docx
pyYAML
docx2pdf
PyPDF2
beautifulsoup4